Use DEVICE_LOCAL memory type for input buffers
On a discrete GPU, this should allow writing directly from the mmap to
the device through the PCIe bus, instead of writing to a memory buffer
and letting the device read it through the PCIe bus. On an integrated
GPU, this change has no effect, since there's only one memory type.