feat: Implement `From<Hash>` for `u64`.
This is an attempt to encode an `Hash` into `u64`. This transformation
will increase the number of collisions as `u64` can represent fewer
values than `[u8; OUT_LEN]` (at the time of writing, `OUT_LEN` is set
to 32).

Bytes from `Hash` are extracted by chunks of 8, and then added to
`u64` with a modular addition.