make g_cpu_features atomic
This change fixes a data race. The race was x86-only and probably
harmless in practice, but as they say, "There's no such thing as a
benign data race." When TSan's happy, everybody's happy.

As part of this change, #ifdef the CPU feature detection code to be
x86-only. This should silence some unused code warnings on other
platforms, and it means that we only depend on stdatomic.h on x86, where
compilers are at least a little bit more likely to be up to date.

Reported by https://github.com/BLAKE3-team/BLAKE3/issues/220.