start working on a refactored assembly implementation
The main goal is to eventually have extended outputs benefit from the
same SIMD optimizations as inputs. To make this easier, I want to factor
out a shared "kernel" routine that can be shared among several
different interfaces:

- compressing chunks
- compressing parents
- producing XOF output
- xor'ing XOF output

The timing here partly coincides with Rust stabilizing inline asm.
That's certainly not necessary for any of this to work, but it gives me
the confidence to try this without needing to master the rules of three
different calling conventions.