Add fn hash_block to guts
provides an efficient way to compute the hash for multiple chunks that do
not start at chunk 0, as part of a larger tree.

This is useful for crates like bao, abao and bao-tree, especially when using
chunk groups.

Adding this required to disable some assertions that contained assumptions
that are no longer true.