Introduce C multi-threading
This is a minimal thread-pool implementation for BLAKE3.

As part of the commit a very simple file mmap-ing example is
introduced that providing file checksuming.

At least on my machines (x86-64 linux boxes) the C implementation
is within a 1% point in performance of the rust b3sum tool.

> $ ls -l ggml-vicuna-7b-1.1-q4_0.bin
-rw-rw-r-- 1 panto panto 3791725184 Ιουν 17 13:34 ggml-vicuna-7b-1.1-q4_0.bin

> $ time b3sum ggml-vicuna-7b-1.1-q4_0.bin
> fe87fecc4427c4661aa46f52a7e286ecd6dd7f5f788564abb2c46d2ed5341584  ggml-vicuna-7b-1.1-q4_0.bin
>
> real	0m0,190s
> user	0m1,583s
> sys	0m0,196s

> $ time ./example-mmap ggml-vicuna-7b-1.1-q4_0.bin
> fe87fecc4427c4661aa46f52a7e286ecd6dd7f5f788564abb2c46d2ed5341584  ggml-vicuna-7b-1.1-q4_0.bin
>
> real	0m0,204s
> user	0m1,525s
> sys	0m0,247s

At my local copy of BLAKE3 with more extensive changes I can get the C
implementation to be slightly faster than the rust one.
But those require more extensive changes which will have to come much
later.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>