Update CMakeLists cmake_minimum_required to suppress warning
cmake 3.31.6 prints a warning when configuring projects
that specify a minimum version of < 3.10. This updates
the version range to suppress the warning.

We chose CMake 3.18 as the upper bound as the vast majority of users
should be using newer CMake versions--CMake 3.18 is shipped in the
current LTS version of Debian (11). Furthermore it precludes some of the
more significant policy changes introduced with CMake 3.19, i.e. we
should get away with not testing the build with every CMake version.