Implement pthread parallelism
This implements a simple version of oneapi::tbb::parallel_invoke using
plain pthreads.

The pthread implementation is slightly less performant, but it may be
better suited for light-weight embedded systems than TBB and C++.

Before bifurcating execution, a thread is allocated from a list of free
threads or it is created if the total number of threads is less than the
number of cores on the system.

If there is no thread available, both the left and right subtrees are
processed.

If a thread could be allocated, the thread is given an execution context
for the left subtree and instructed to start execution via a barrier.

The current thread continues to process the right subtree and
synchronises again with the allocated thread via a barrier and returns
it to the list of free threads.